requirements.txt
streamlit==1.31.0
google-generativeai==0.3.0
PyPDF2==3.0.1
Pillow==10.1.0
python-dotenv==1.0.0
pdf2image==1.16.3
numpy==1.24.3
pandas==2.1.0
